requirements.txt
azure.identity
azure.mgmt.costmanagement